+++ packages/slog_basic/CHANGELOG.md
@@ -4,4 +4,8 @@
 - feat: added json printer & interceptor printer
 
 # 0.1.0+2
-- added missing export
+- bug: added missing export
+
+# 0.1.1
+- bug: fixes json printer error for some types
+- dep: ups version of slog_core