+++ packages/slog_core/CHANGELOG.md
@@ -3,3 +3,8 @@
 - feat: initial commit 🎉
 - feat: added log interceptor & printer
 - feat: added default logger
+
+# 0.2.0
+
+- feat: allows running the logger to make use of an [Isolate](https://dart.dev/guides/language/concurrency) 
+- feat: adds example for basic usage